refactor: move analysis out of datapipe
