+++ cmd/api/src/analysis/analysis.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package datapipe
+package analysis
 
 import (
 	"context"
